+++ src/tools/Demultiplexing/DemultiplexedLinesWriter.hpp
@@ -13,9 +13,9 @@ class DemultiplexedLinesWriter : private Mapping<MappingPolicy, FilePolicy>
 
         void demultiplex_wrapper(const std::string& line,
                                 const input& input,
-                                std::atomic<int>& lineCount,
-                                const int& totalReadCount,
-                                std::atomic<int>& elementsInQueue);
+                                std::atomic<unsigned long long>& lineCount,
+                                const unsigned long long& totalReadCount,
+                                std::atomic<long long int>& elementsInQueue);
         void initialize_output_files(const input& input,const std::vector<std::pair<std::string, char> >& patterns);
         void run_mapping(const input& input);
